+++ pkg/server/app/models.go
@@ -31,7 +31,7 @@ type App struct {
 	Name        string     `json:"name"`
 	Team        string     `json:"-"`
 	ProcessType string     `json:"processType"`
-	Host        string     `json:"host"`
+	VirtualHost string     `json:"virtualHost"`
 	Limits      *Limits    `json:"-"`
 	Autoscale   *Autoscale `json:"-"`
 	EnvVars     []*EnvVar  `json:"envVars"`
@@ -125,7 +125,7 @@ func newApp(req *appb.CreateRequest) *App {
 		},
 		Name:        req.Name,
 		ProcessType: processType,
-		Host:        req.Host,
+		VirtualHost: req.VirtualHost,
 		Team:        req.Team,
 		EnvVars:     []*EnvVar{},
 	}