+++ pkg/server/deploy/deploy.go
@@ -28,7 +28,7 @@ type Operations interface {
 type K8sOperations interface {
 	PodRun(podSpec *PodSpec) (io.ReadCloser, <-chan int, error)
 	CreateOrUpdateDeploy(deploySpec *DeploySpec) error
-	ExposeApp(namespace, name, host string, w io.Writer) error
+	ExposeApp(namespace, name, vHost string, w io.Writer) error
 	ReplicaSetListByLabel(namespace, label, value string) ([]*ReplicaSetListItem, error)
 	DeployRollbackToRevision(namespace, name, revision string) error
 }
@@ -116,7 +116,7 @@ func (ops *DeployOperations) exposeService(a *app.App, w io.Writer) error {
 	if a.ProcessType != app.ProcessTypeWeb {
 		return nil
 	}
-	if err := ops.k8s.ExposeApp(a.Name, a.Name, a.Host, w); err != nil {
+	if err := ops.k8s.ExposeApp(a.Name, a.Name, a.VirtualHost, w); err != nil {
 		return err
 	}
 	return nil // already exposed